requirements.txt
beautifulsoup4
playwright
pyyaml
requests
tinydb
python-dotenv
langchain-google-genai
langchain-openai
langchain
pydantic